+++ internal/scanner/sast/testdata/xss.js
@@ -1,0 +1,2 @@
+const x = eval(userInput);
+element.innerHTML = request.body;